Moving some bits around, experimenting with URI generation
